+++ StripePayments/StripePayments/Source/PaymentHandler/STPPaymentHandler.swift
@@ -873,7 +873,8 @@ public class STPPaymentHandler: NSObject {
             .bizum,
             .krCard,
             .naverPay,
-            .payco:
+            .payco,
+            .sequra:
             return false
 
         case .unknown: